+++ ch08/user-management-system/src/App.vue
@@ -5,5 +5,5 @@
 </template>
 
 <script setup>
-import Layout from "./components/Layout.vue";
+import Layout from "./layout/Layout.vue";
 </script>
